comment functions

+++ srcs/ft_lst_/list/srcs/list_size.c
@@ -1,6 +1,6 @@
 #include "list.h"
 
-int	list_size(t_list *head)
+size_t	list_size(t_list *head)
 {
 	t_list	*pos;
 	int		ret;
